+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 
 - `Subscription.subscribeToChatMessages`
 - `Subscription.subscribeToChatOnlineStatuses`
+- `Subscription.subscribeToChatTypingStatuses`
 
 ## [0.19.0](https://github.com/neelkamath/omni-chat-backend/releases/tag/v0.19.0) - 2021-05-21